+++ WB-weather/providers/WB-newProviderTemplate.js
@@ -1,0 +1,8 @@
+/* Magic Mirror - WallberryTheme <3
+ * Module: WB-weather
+ *
+ * By JSC (@delightedCrow)
+ * MIT Licensed.
+ */
+
+ // coming soon :)